+++ openfisca_core/base_functions.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 
-
 import numpy as np
 
 from . import periods
@@ -8,11 +7,9 @@ from . import periods
 
 def permanent_default_value(formula, simulation, period, *extra_params):
     if formula.function is not None:
-        return formula.function(simulation, period, *extra_params)
+        return formula.exec_function(simulation, period, *extra_params)
     holder = formula.holder
-    column = holder.column
-    array = np.empty(holder.entity.count, dtype = column.dtype)
-    array.fill(column.default)
+    array = holder.default_array()
     return period, array
 
 
@@ -49,27 +46,22 @@ def requested_period_added_value(formula, simulation, period, *extra_params):
             if array is not None:
                 return period, array
     if formula.function is not None:
-        return formula.function(simulation, period, *extra_params)
-    array = np.empty(holder.entity.count, dtype = column.dtype)
-    array.fill(column.default)
+        return formula.exec_function(simulation, period, *extra_params)
+    array = holder.default_array()
     return period, array
 
 
 def requested_period_default_value(formula, simulation, period, *extra_params):
     if formula.function is not None:
-        return formula.function(simulation, period, *extra_params)
+        return formula.exec_function(simulation, period, *extra_params)
     holder = formula.holder
-    column = holder.column
-    array = np.empty(holder.entity.count, dtype = column.dtype)
-    array.fill(column.default)
+    array = holder.default_array()
     return period, array
 
 
 def requested_period_default_value_neutralized(formula, simulation, period, *extra_params):
     holder = formula.holder
-    column = holder.column
-    array = np.empty(holder.entity.count, dtype = column.dtype)
-    array.fill(column.default)
+    array = holder.default_array()
     return period, array
 
 
@@ -90,10 +82,8 @@ def requested_period_last_value(formula, simulation, period, *extra_params, **kw
             next_period, next_array = known_values[-1]
             return period, last_result
     if formula.function is not None:
-        return formula.function(simulation, period, *extra_params)
-    column = holder.column
-    array = np.empty(holder.entity.count, dtype = column.dtype)
-    array.fill(column.default)
+        return formula.exec_function(simulation, period, *extra_params)
+    array = holder.default_array()
     return period, array
 
 
@@ -116,8 +106,6 @@ def last_duration_last_value(formula, simulation, period, *extra_params):
                 elif last_result.get(extra_params):
                     return output_period, last_result.get(extra_params)
     if formula.function is not None:
-        return formula.function(simulation, period, *extra_params)
-    column = holder.column
-    array = np.empty(holder.entity.count, dtype = column.dtype)
-    array.fill(column.default)
+        return formula.exec_function(simulation, period, *extra_params)
+    array = holder.default_array()
     return period, array