+++ openfisca_core/columns.py
@@ -16,6 +16,7 @@ from .enumerations import Enum
 def N_(message):
     return message
 
+
 year_or_month_or_day_re = re.compile(ur'(18|19|20)\d{2}(-(0?[1-9]|1[0-2])(-([0-2]?\d|3[0-1]))?)?$')
 
 
@@ -27,9 +28,7 @@ class Column(object):
     default = 0
     dtype = float
     end = None
-    entity = None  # Obsolete: To remove once build_..._couple() functions are no more used.
-    entity_key_plural = None
-    entity_class = None
+    entity = None
     formula_class = None
     is_period_size_independent = False  # When True, value of column doesn't depend from size of period (example: age)
     is_permanent = False  # When True, value of column doesn't depend from time (example: ID, birth)
@@ -52,7 +51,6 @@ class Column(object):
             self.default = default
         if end is not None:
             self.end = end
-        self.entity = entity or 'ind'
         if function is not None:
             self.function = function
         if is_permanent:
@@ -145,7 +143,7 @@ class Column(object):
                 end = end.isoformat()
             self_json['end'] = end
         if self.entity is not None:
-            self_json['entity'] = self.entity_key_plural
+            self_json['entity'] = self.entity.key
         if self.label is not None:
             self_json['label'] = self.label
         start_line_number = self.formula_class.start_line_number