+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## 4.0.0-Beta 1
+
+  * Refactor formula syntax
+
+    - Don't use the object simulation in formula writing. The entity the variable is calculated for is now the first argument of the formula, for instance `person` or `family`.
+
+    - The `legislation` is now the (optional) third argument of a formula. It is a function that can be called with a period (or an instant) for argument, and will compute the legislation at `period.start`.
+
+    -  Don't use `calculate_add` and equivalents. Instead, `add` and `divide` are just given in the `options` argument, for instance `options = [ADD]`.
+
+    - Rename `entity_class` to `entity`.
+
+    - Don't explicitly use a `calculate` method, but the `person('salary', period)` notation instead.
+
+  * Introduce implicit conversions between entities, and improve aggregations formulas.
+
+  * Deprecate PersonToEntityColumn and EntityToPersonColumn
+
+  * Change the way entities are declared
+
+  More information on https://github.com/openfisca/openfisca-core/pull/415
+
 ## 3.1.0
 
 * Add a `DeprecationWarning` when using a `DateCol` with no `default`, but keep the default date to 1970-01-01.